Use time.perf_counter for DNB rate limiting
